requirements.txt
certifi==2018.8.24
chardet==3.0.4
idna==2.7
lxml==4.2.4
PyYAML==3.13
requests==2.19.1
requests-file==1.4.3
tldextract==2.2.0
urllib3==1.23
